Release v1.3.1

+++ Projects/NUCLEO-WBA55CG/Applications/BLE/BLE_p2pServer/System/Modules/ble_timer.c
@@ -47,7 +47,7 @@ static BLE_TIMER_t* BLE_TIMER_timer;
 /* Private functions prototype------------------------------------------------*/
 void BLE_TIMER_Background(void);
 static void BLE_TIMER_Callback(void* arg);
-static BLE_TIMER_t* BLE_TIMER_GetFromList(tListNode * listHead, uint8_t id);
+static BLE_TIMER_t* BLE_TIMER_GetFromList(tListNode * listHead, uint16_t id);
 
 void BLE_TIMER_Init(void)
 {
@@ -61,7 +61,7 @@ void BLE_TIMER_Init(void)
   UTIL_TIMER_Init();
 }
 
-uint8_t BLE_TIMER_Start(uint8_t id, uint32_t timeout)
+uint8_t BLE_TIMER_Start(uint16_t id, uint32_t timeout)
 {
   /* If the timer's id already exists, stop it */
   BLE_TIMER_Stop(id);
@@ -97,7 +97,7 @@ uint8_t BLE_TIMER_Start(uint8_t id, uint32_t timeout)
   return BLE_STATUS_SUCCESS;
 }
 
-void BLE_TIMER_Stop(uint8_t id){
+void BLE_TIMER_Stop(uint16_t id){
   /* Search for the id in the timers list */
   BLE_TIMER_t* timer = BLE_TIMER_GetFromList(&BLE_TIMER_List, id);
 
@@ -113,7 +113,7 @@ void BLE_TIMER_Stop(uint8_t id){
 
 void BLE_TIMER_Background(void)
 {
-  BLEPLATCB_TimerExpiry( (uint8_t)BLE_TIMER_timer->id);
+  BLEPLATCB_TimerExpiry( (uint16_t)BLE_TIMER_timer->id);
   HostStack_Process( );
 
   /* Delete the BLE_TIMER_timer from the list */
@@ -129,7 +129,7 @@ static void BLE_TIMER_Callback(void* arg)
   UTIL_SEQ_SetTask( 1U << CFG_TASK_BLE_TIMER_BCKGND, CFG_SEQ_PRIO_0);
 }
 
-static BLE_TIMER_t* BLE_TIMER_GetFromList(tListNode * listHead, uint8_t id)
+static BLE_TIMER_t* BLE_TIMER_GetFromList(tListNode * listHead, uint16_t id)
 {
   BLE_TIMER_t* currentNode = (BLE_TIMER_t*)listHead->next;
   while((tListNode *)currentNode != listHead)
