+++ Projects/STM32WBA55G-DK1/Applications/BLE/BLE_Audio_PBP_Sink/System/Modules/ble_timer.h
@@ -7,7 +7,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2022 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -23,13 +23,13 @@
 
 void BLE_TIMER_Init( void );
 
-uint8_t BLE_TIMER_Start( uint8_t id,
+uint8_t BLE_TIMER_Start( uint16_t id,
                      uint32_t ms_timeout );
 
-void BLE_TIMER_Stop( uint8_t id );
+void BLE_TIMER_Stop( uint16_t id );
 
 /* Callback
  */
-void BLE_TIMERCB_Expiry( uint8_t id );
+void BLE_TIMERCB_Expiry( uint16_t id );
 
 #endif /* BLE_TIMER_H__ */
