+++ Projects/STM32WBA55G-DK1/Applications/BLE/BLE_HID_Mouse/System/Modules/ble_timer.h
@@ -23,13 +23,13 @@
 
 void BLE_TIMER_Init( void );
 
-uint8_t BLE_TIMER_Start( uint8_t id,
+uint8_t BLE_TIMER_Start( uint16_t id,
                      uint32_t ms_timeout );
 
-void BLE_TIMER_Stop( uint8_t id );
+void BLE_TIMER_Stop( uint16_t id );
 
 /* Callback
  */
-void BLE_TIMERCB_Expiry( uint8_t id );
+void BLE_TIMERCB_Expiry( uint16_t id );
 
 #endif /* BLE_TIMER_H__ */
